MC-37605: Validate Magento functionality with RO pub/static directory
